+++ src/assets/index.ts
@@ -1,0 +1,8 @@
+export { default as landingKeyboard } from "./images/landingKeyboard.webp";
+export { default as topSectionItem1 } from "./images/topSectionItem1.webp";
+export { default as landingWrite } from "./images/landingWrite.webp";
+export { default as middleSectionItem1 } from "./images/middleSectionItem1.webp";
+export { default as bottomSectionItem1 } from "./images/bottomSectionItem1.webp";
+export { default as bottomSectionItem2 } from "./images/bottomSectionItem2.webp";
+export { default as bottomSectionItem3 } from "./images/bottomSectionItem3.webp";
+export { default as backgroundImage } from "./images/topSectionBackgroundImg.webp";
